pyproject.toml
[project]
name = "arxiv-podcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.33.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
]
